多态

+++ 6.22作业/Shape/src/club/banyuan/Rectangle.java
@@ -2,10 +2,10 @@ package club.banyuan;
 
 public class Rectangle extends Shape {
 
-  private int length;
-  private int width;
+  private double length;
+  private double width;
 
-  public Rectangle(int length, int width) {
+  public Rectangle(double length, double width) {
     this.length = length;
     this.width = width;
   }
